requirements.txt
discord.py==0.16.12
requests==2.18.4
discord==0.0.2
python-dotenv==0.7.1
